Replaces the default astronomer image with 2.0.0-buster-onbuild (#401)
* Replaces the default astronomer image with 2.0.0-buster-onbuild

The image with the tag "latest_onbuild" references image b17129847bec, which is an alpine image for Airflow 1.10.7. This image is broken. An error is received when the user attempts to rebuild (aka deploy their dags to Astronomer Platform). There is an issue to implement a proper fix (https://github.com/astronomer/issues/issues/2275), but until that can be done, this PR changes the default to a 2.0 on-build debian image. This ensures that users who are using Astro cli for the first time to create a new deployment project will start with a non-broken image, and hopefully have a smoother experience.

* tweak config

Co-authored-by: Andrii Soldatenko <andrii.soldatenko@gmail.com>