Add flow command to astro-cli
Create a docker file Dockerfile.sql_cli with astro-sql-cli python
package installed in it. Add a `flow` command to astro-cli which
calls the docker image built from the above docker file. Flow
commands are passed from the astro-cli `flow` command to the docker
container run with the image.

Define a config SQLCLI which has a boolean to  enable/disable `flow`
usage.