Fix advanced example DAG and modernize
The current advanced example DAG has a bug which causes task failures when executed on Saturdays. Also, the DAG wouldn't be able to run out of the box due to the use of the EmailOperator and it's required SMTP configurations to execute successfully.

This PR fixes said Saturday-execution bug, replaces the EmailOperator with another TaskFlow function, and also modernizes the DAG to use TaskFlow API equivalents and syntax that's available with Airflow 2.5.1.