Bulk convert to using testify/suite for tests (#1185)
There were a couple of places where `cmdExec` or similar was changed but not
changed back.

Now by this being a test suite it's handled for us.

As part of this work a number of other "leaking" side-effects
in tests are cleaned up properly. However I
wasn't able
to fix all of them, so those packages are left as stdlib testing only.

I also didn't convert everything we could have to use Setup functions.